refactor : Product Domain Context > Product - 패키지 생성 및 기존 클래스 이동 - JPARepository로 교체
