+++ src/test/java/kitchenpos/acceptance/MenuAcceptanceTest.java
@@ -10,7 +10,7 @@ import kitchenpos.acceptance.helper.KitchenPosBehaviors;
 import kitchenpos.domain.Menu;
 import kitchenpos.menu.domain.MenuGroup;
 import kitchenpos.domain.MenuProduct;
-import kitchenpos.domain.Product;
+import kitchenpos.product.domain.Product;
 import kitchenpos.fixture.MenuFixtureFactory;
 import kitchenpos.fixture.MenuProductFixtureFactory;
 import org.assertj.core.util.Lists;
