+++ src/test/java/kitchenpos/fixture/ProductFixtureFactory.java
@@ -1,23 +1,14 @@
 package kitchenpos.fixture;
 
 import java.math.BigDecimal;
-import kitchenpos.domain.Product;
+import kitchenpos.product.domain.Product;
 
 public class ProductFixtureFactory {
     private ProductFixtureFactory() {
     }
 
     public static Product createProduct(String name, int price) {
-        Product product = new Product();
-        product.setName(name);
-        product.setPrice(new BigDecimal(price));
-        return product;
+        return new Product(name, new BigDecimal(price));
     }
 
-    public static Product createProduct(String name, long price) {
-        Product product = new Product();
-        product.setName(name);
-        product.setPrice(new BigDecimal(price));
-        return product;
-    }
 }
